finance_analyst_api/requirement.txt
djangorestframework
markdown
django-filter
psycopg2
django-environ